Update parent-st-anything-ethernet.groovy
Changed over the default sendEvent routine for sending data to child events over to executing a generateEvent routine.  This should allow for updates to the child devices when data is received and allow future customization.
NOTE: I have only tested this on contact, temperature, and humidity which all work properly.  It should work on all child devices the same way once they are updated but testing, especially with the door control, needs to be done.