Update SmartThingsESP32WiFi.cpp
So under certain conditions the ESP32 will not reconnect and there doesn't appear to be a fix yet other then rebooting the controller (which works).  The WiFi.reconnect gets called but it can't actually reconnect.  So the above change is a band-aid for those having this issue like me.  It shouldn't affect a properly working ESP32.