Update parent-st-anything-ethernet.groovy
Think the biggest issue I and others have had is when first configuring the parent and not using any buttons is errors about it being required.  The only solution I have seen is to make it 1, or anything other then 0, save it, then go back and change it to 0.  I think based on testing this is because of the default value: 0 part not fully working (on SmartThings side).  So I'm suggesting keeping it as required but removing the default value.  It should get around this issue.  It will still force the option to be set but it shouldn't give a error when its set to 0 the first time.