Create EX_RGB_Dim.h
This is my first attempt at doing this (other then the Adafruit color detector which was really jsut cut and paste).  I'm attempt to make the library universal for both a regular Arduino/ESP8266 and the ESP32 which uses a LedC library and PWM channels.  The idea would be the device sets the 3 pins (R,G,B) and for a ESP32 also the 3 channels (R,G,B).  There is also a commonAnode boolean that can be flipped depending on the LED you are using.  I.e. this could be used for a stand alone RGB LED or RGB LED strips with common anode or cathode.

The only two things passed to it would be On/Off and HEX of the color (standard #FFFFFF format).  One its working the code could be modified into a RGBW with only the addition of another pin and channel since you can send the HEX as #FFFFFFFF.  

Please look this over and see if it makes sense and fits in with other libraries. I used the EX_Switch_Dim as the base since that made the most sense.  If this is incorrect please let me know what to change or tweak it.  Like I;m not sure on the "sets" section if you would just have a "setRGBPins" that takes all three pins as inputs, sane with the channels.  Once its "correct" I'll see if I can figure out the actual code.  I have it in Arduino already and it works great....just have to figure out how to convert it.

Oh and here is what I think would be the config from the sketch file:

 //Executors
  static st::EX_RGB_Dim         executor1(F("rgbSwitch1"), PIN_RGB1_Red, PIN_RGB1_Green, PIN_RGB1_Blue, true, 0, 1, 2);