Update EX_RGB_Dim.h
Since it should get on/off commands there should be a writeStateToDevice function.  In the EX_Switch_Dim this turns the actual output pin on and off.  Here it will change the state of the device but o a "off" it will write 0's to the PWM pins (#000000) and for a "on" it would send the last hex value through which it would still have locally although I am also sending t from the DTH just in case.