Update ST_Anything_Multiples_ESP32WiFi.ino
Updated the example sketch for the RGB LED/Strip capability.  Removed a water, illuminance, and a motion sensor to fit in the three pins needed by the RGB LED or strip.  Believe I reordered all the numbering correctly.

Once this is merged in and I get more people to test it I'll try to update the 8266 example sketch.  Not sure if any others really need to be done...think 8266 is the most used and the 32 will be as time goes on.