Create PS_Generic.cpp
Let me know what you think of something like this.  I think it would give the framework to people that are a little confused with how it works more information to expand on their own.  It would also be a good framework for any type of custom sensor (like the Adafruit TCS color/light sensor) as they could just copy and paste the example code into the .cpp file without having to really modify anything else.

If it makes sense and you want to commit it I'll PR the changes to the parent and a DTH to support it.