Update EX_Servo.cpp
Enables users to change endpoints associated with level 0 and level 100 by swapping values for the min level angle argument with the max level angle argument in the sketch. The min level angle can be greater than the max level angle.  This provides more flexibility in connecting a servo to mechanical hardware.